+++ mayan/apps/django_gpg/api_views.py
@@ -17,6 +17,11 @@ class APIKeyListView(generics.ListCreateAPIView):
     queryset = Key.objects.all()
     serializer_class = KeySerializer
 
+    def get_instance_extra_data(self):
+        return {
+            '_event_actor': self.request.user
+        }
+
 
 class APIKeyView(generics.RetrieveDestroyAPIView):
     """