+++ mayan/apps/django_gpg/tests/test_api.py
@@ -2,6 +2,7 @@ from rest_framework import status
 
 from mayan.apps.rest_api.tests.base import BaseAPITestCase
 
+from ..events import event_key_created
 from ..models import Key
 from ..permissions import (
     permission_key_delete, permission_key_upload, permission_key_view
@@ -12,14 +13,23 @@ from .mixins import KeyAPIViewTestMixin, KeyTestMixin
 
 
 class KeyAPITestCase(KeyTestMixin, KeyAPIViewTestMixin, BaseAPITestCase):
+    _test_event_object_name = 'test_key_private'
+
     def test_key_create_api_view_no_permission(self):
+        self._clear_events()
+
         response = self._request_test_key_create_api_view()
         self.assertEqual(response.status_code, status.HTTP_403_FORBIDDEN)
 
         self.assertEqual(Key.objects.all().count(), 0)
 
+        event = self._get_test_object_event()
+        self.assertEqual(event, None)
+
     def test_key_create_api_view_with_permission(self):
         self.grant_permission(permission=permission_key_upload)
+
+        self._clear_events()
 
         response = self._request_test_key_create_api_view()
         self.assertEqual(response.status_code, status.HTTP_201_CREATED)
@@ -31,13 +41,23 @@ class KeyAPITestCase(KeyTestMixin, KeyAPIViewTestMixin, BaseAPITestCase):
         self.assertEqual(Key.objects.count(), 1)
         self.assertEqual(key.fingerprint, TEST_KEY_PRIVATE_FINGERPRINT)
 
+        event = self._get_test_object_event()
+        self.assertEqual(event.actor, self._test_case_user)
+        self.assertEqual(event.target, self.test_key_private)
+        self.assertEqual(event.verb, event_key_created.id)
+
     def test_key_delete_api_view_no_permission(self):
         self._create_test_key_private()
+
+        self._clear_events()
 
         response = self._request_test_key_delete_api_view()
         self.assertEqual(response.status_code, status.HTTP_404_NOT_FOUND)
 
         self.assertEqual(Key.objects.count(), 1)
+
+        event = self._get_test_object_event()
+        self.assertEqual(event, None)
 
     def test_key_delete_api_view_with_access(self):
         self._create_test_key_private()
@@ -45,16 +65,26 @@ class KeyAPITestCase(KeyTestMixin, KeyAPIViewTestMixin, BaseAPITestCase):
             obj=self.test_key_private, permission=permission_key_delete
         )
 
+        self._clear_events()
+
         response = self._request_test_key_delete_api_view()
         self.assertEqual(response.status_code, status.HTTP_204_NO_CONTENT)
 
         self.assertEqual(Key.objects.count(), 0)
 
+        event = self._get_test_object_event()
+        self.assertEqual(event, None)
+
     def test_key_detail_api_view_no_permission(self):
         self._create_test_key_private()
 
+        self._clear_events()
+
         response = self._request_test_key_detail_api_view()
         self.assertEqual(response.status_code, status.HTTP_404_NOT_FOUND)
+
+        event = self._get_test_object_event()
+        self.assertEqual(event, None)
 
     def test_key_detail_api_view_with_access(self):
         self._create_test_key_private()
@@ -62,8 +92,13 @@ class KeyAPITestCase(KeyTestMixin, KeyAPIViewTestMixin, BaseAPITestCase):
             obj=self.test_key_private, permission=permission_key_view
         )
 
+        self._clear_events()
+
         response = self._request_test_key_detail_api_view()
         self.assertEqual(response.status_code, status.HTTP_200_OK)
         self.assertEqual(
             response.data['fingerprint'], self.test_key_private.fingerprint
         )
+
+        event = self._get_test_object_event()
+        self.assertEqual(event, None)